fix: clippy clean under rust 1.98 (unnecessary_cast, field_reassign_with_default)

+++ src/evolve/analysis.rs
@@ -735,7 +735,7 @@ mod tests {
             .iter()
             .find(|p| p.pattern_type == "long_debug_loop")
             .unwrap();
-        assert_eq!(loop_pat.count, runs as u64);
+        assert_eq!(loop_pat.count, runs);
     }
 
     #[test]
